Push namespace transformer code to plugin.
